putting resolution proofs utils in its own module
